+++ grails-app/migrations/changelog.groovy
@@ -37,4 +37,5 @@ databaseChangeLog = {
 	include file: 'big-decimal.xml'
     include file: 'modules-uppercase.xml'
     include file: 'rename-service.xml'
+    include file: 'assessments.xml'
 }